Added Countdown menu, serverinfo status updates
A countdown menu will now replace the "10... 9..." messages which are
displayed during the last 10 seconds of pm.

A serverinfo qwtfver VERSION field is now added
TF version updated to 2.92a
The highest team score is now displayed in red in the match results output
Misc small alignment changes.

Hopefully there are no bugs :F
Thanks to Spike and Jogi for the help.