List and newcountdown changes
Added /list to see who's not ready. Works once every 10 seconds.

Added some code for the newcountdown menu. The main code in clan.qc
isn't updated yet, just the checks to see if status bar is on and what
the res is.

Fixed a sbar bug for spectators in spect.qc where setinfo sbr wasn't
checked.