Huge status bar update, code cleanup
Removed a lot of unnecessary code and updated the status bar in the mean
time.

Added some color to it, Spy shows "not undercover" when he's not
undercover, and probably fixed a lot of bugs as a result of
spacing/alignment. Other misc status bar changes, read the next
TF-Versions.txt

Down to 36 warnings